refactor(ros2agnocast): reorganize verb directory to contain only ros2 commands
- Move helper modules (_a2r_bridge_activator, _bridged_ros2cli) from
  verb/ to the top-level ros2agnocast package
- Rename verb files to follow <ros2_verb>_<sub_verb> convention:
  - bridge_daemon_status.py    -> agnocast_bridge_daemon_status.py
  - discovery_daemon_status.py -> agnocast_discovery_daemon_status.py
  - generate_bridge_plugins.py -> agnocast_generate_bridge_plugins.py
  - list_agnocast.py           -> topic_list_agnocast.py
  - version.py                 -> agnocast_version.py
- Update all import paths and setup.py entry points accordingly
- Update patch paths in test_a2r_bridge_activator.py
